requirements.txt
runpod>=1.7.0
chatterbox-tts>=0.1.0
torch>=2.1.0
torchaudio>=2.1.0
pydub>=0.25.0
